align s6/etcd versions in baseos and baseos-alpine
This matters because earlier versions of s6, such as the one used by
baseos do not respect the S6_KEEP_ENV setting. We need this setting to
ensure environment variables set on the container are visible to our run
scripts. They have a with-contenv that is supposed to restore
environment variables (without S6_KEEP_ENV) but this does not restore
null/empty environment variables.